refactor: TemplateType.php -> ActionType.php
